pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-jwt-extended>=4.7.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-mail>=0.10.0",
    "wtforms>=3.2.1",
    "werkzeug>=3.1.3",
    "numpy>=2.3.1",
    "pillow>=11.3.0",
    "matplotlib>=3.10.3",
    "pdfplumber>=0.11.7",
    "moviepy>=2.2.1",
    "imageio>=2.37.0",
    "imageio-ffmpeg>=0.6.0",
    "sqlalchemy>=2.0.41",
    "flask-wtf>=1.2.2",
    "flask-login>=0.6.3",
]
